+++ docs/demo/package.json
@@ -11,7 +11,7 @@
     "preview": "astro preview"
   },
   "dependencies": {
-    "@tutorialkit/components-react": "workspace:*",
+    "@tutorialkit/react": "workspace:*",
     "react": "^18.3.1",
     "react-dom": "^18.3.1"
   },
