+++ docs/tutorialkit.dev/src/content/docs/reference/react-components.mdx
@@ -32,17 +32,17 @@ These components use TutorialKit's design system which is based on an atomic CSS
 <PackageManagerTabs>
   <Fragment slot="npm">
   ```shell
-  npm install @tutorialkit/components-react @tutorialkit/theme
+  npm install @tutorialkit/react @tutorialkit/theme
   ```
   </Fragment>
   <Fragment slot="pnpm">
   ```shell
-  pnpm install @tutorialkit/components-react @tutorialkit/theme
+  pnpm install @tutorialkit/react @tutorialkit/theme
   ```
   </Fragment>
   <Fragment slot="yarn">
   ```shell
-  yarn add @tutorialkit/components-react @tutorialkit/theme
+  yarn add @tutorialkit/react @tutorialkit/theme
   ```
   </Fragment>
 </PackageManagerTabs>
