+++ e2e/package.json
@@ -14,7 +14,7 @@
     "@iconify-json/svg-spinners": "^1.1.2",
     "@playwright/test": "^1.46.0",
     "@tutorialkit/astro": "workspace:*",
-    "@tutorialkit/components-react": "workspace:*",
+    "@tutorialkit/react": "workspace:*",
     "@tutorialkit/runtime": "workspace:*",
     "@tutorialkit/theme": "workspace:*",
     "@tutorialkit/types": "workspace:*",
