+++ packages/astro/README.md
@@ -2,7 +2,7 @@
 
 This **[Astro integration][astro-integration]** adds [TutorialKit](https://tutorialkit.dev/) to your project so that you can use TutorialKit's tutorial format for your astro content.
 
-This integration adds routes to serve your tutorial. It uses `@tutorialkit/components-react` for the dynamic part of the experience.
+This integration adds routes to serve your tutorial. It uses `@tutorialkit/react` for the dynamic part of the experience.
 
 ## License
