+++ packages/astro/src/default/components/NavWrapper.tsx
@@ -1,4 +1,4 @@
-import { Nav } from '@tutorialkit/components-react';
+import { Nav } from '@tutorialkit/react';
 import type { Lesson, NavList } from '@tutorialkit/types';
 
 interface Props {
