+++ packages/react/scripts/build.js
