+++ packages/react/src/BootScreen.tsx
