+++ packages/react/src/Panels/TerminalPanel.tsx
