+++ packages/react/src/Panels/WorkspacePanel.tsx
