+++ packages/react/src/core.ts
