+++ packages/react/src/core/CodeMirrorEditor/BinaryContent.tsx
