+++ packages/react/src/core/FileTree.tsx
