+++ packages/react/src/core/Terminal/index.tsx
