+++ packages/react/src/core/Terminal/theme.ts
