+++ packages/react/src/styles/resize-panel.module.css
