+++ packages/react/src/styles/terminal.css
