+++ packages/react/src/utils/mobile.ts
