+++ pnpm-workspace.yaml
@@ -1,6 +1,5 @@
 packages:
   - 'packages/*'
-  - 'packages/components/*'
   - 'docs/*'
   - 'extensions/*'
   - 'integration'
